packages/orly/generate/requirements.txt
Pillow==10.0.0
FontTools==4.41.1
